Shifting swagger related method to existing swagger file. Shifting telemetry methods to its own file
