+++ pkg/gofr/gofr.go
@@ -119,19 +119,6 @@ func New() *App {
 	return app
 }
 
-func (a *App) checkAndAddOpenAPIDocumentation() {
-	// If the openapi.json file exists in the static directory, set up routes for OpenAPI and Swagger documentation.
-	if _, err := os.Stat("./static/" + gofrHTTP.DefaultSwaggerFileName); err == nil {
-		// Route to serve the OpenAPI JSON specification file.
-		a.add(http.MethodGet, "/.well-known/"+gofrHTTP.DefaultSwaggerFileName, OpenAPIHandler)
-		// Route to serve the Swagger UI, providing a user interface for the API documentation.
-		a.add(http.MethodGet, "/.well-known/swagger", SwaggerUIHandler)
-		// Catchall route: any request to /.well-known/{name} (e.g., /.well-known/other)
-		// will be handled by the SwaggerUIHandler, serving the Swagger UI.
-		a.add(http.MethodGet, "/.well-known/{name}", SwaggerUIHandler)
-	}
-}
-
 // NewCMD creates a command-line application.
 func NewCMD() *App {
 	app := &App{}
@@ -222,37 +209,6 @@ func (a *App) Run() {
 	}()
 
 	wg.Wait()
-}
-
-func (a *App) hasTelemetry() bool {
-	return a.Config.GetOrDefault("GOFR_TELEMETRY", defaultTelemetry) == "true"
-}
-
-func (a *App) sendTelemetry(client *http.Client, isStart bool) {
-	url := fmt.Sprint(gofrHost, shutServerPing)
-
-	if isStart {
-		url = fmt.Sprint(gofrHost, startServerPing)
-
-		a.container.Info("GoFr records the number of active servers. Set GOFR_TELEMETRY=false in configs to disable it.")
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), pingTimeout)
-	defer cancel()
-
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, url, http.NoBody)
-	if err != nil {
-		return
-	}
-
-	req.Header.Set("Connection", "close")
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return
-	}
-
-	resp.Body.Close()
 }
 
 // Shutdown stops the service(s) and close the application.
